backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0.0
pony>=0.7.17
python-decouple>=3.8
PyJWT>=2.8.0
anthropic>=0.40.0
